+++ kfinance/domains/competitors/competitor_tools.py
@@ -10,7 +10,7 @@ from kfinance.domains.competitors.competitor_models import CompetitorResponse, C
 from kfinance.integrations.tool_calling.tool_calling_models import (
     KfinanceTool,
     ToolArgsWithIdentifiers,
-    ToolRespWithErrors,
+    ToolRespWithIdInfoAndErrors,
 )
 
 
@@ -19,8 +19,8 @@ class GetCompetitorsFromIdentifiersArgs(ToolArgsWithIdentifiers):
     competitor_source: CompetitorSource
 
 
-class GetCompetitorsFromIdentifiersResp(ToolRespWithErrors):
-    results: dict[str, CompetitorResponse]
+class GetCompetitorsFromIdentifiersResp(ToolRespWithIdInfoAndErrors[CompetitorResponse]):
+    pass
 
 
 class GetCompetitorsFromIdentifiers(KfinanceTool):
@@ -66,8 +66,15 @@ async def get_competitors_from_identifiers(
     {
         "results": {
             "SPGI": {
-                {'company_id': "C_35352", 'company_name': 'The Descartes Systems Group Inc.'},
-                {'company_id': "C_4003514", 'company_name': 'London Stock Exchange Group plc'}
+                'company_name': 'S&P Global Inc.',
+                'ticker': 'NYSE:SPGI',
+                'country': 'USA',
+                'data': {
+                    'competitors': [
+                        {'company_id': "C_35352", 'company_name': 'The Descartes Systems Group Inc.'},
+                        {'company_id': "C_4003514", 'company_name': 'London Stock Exchange Group plc'}
+                    ]
+                }
             }
         },
         'errors': ['No identification triple found for the provided identifier: NON-EXISTENT of type: ticker']
@@ -101,7 +108,11 @@ async def get_competitors_from_identifiers(
         else:
             results[task.result_key] = task.result
 
-    resp_model = GetCompetitorsFromIdentifiersResp(results=results, errors=errors)
+    resp_model = GetCompetitorsFromIdentifiersResp(
+        identifier_results=results,
+        identifier_info=id_triple_resp.identifiers_to_id_triples,
+        errors=errors,
+    )
     return resp_model