+++ kfinance/domains/statements/tests/test_statement_tools.py
@@ -3,7 +3,7 @@ import pytest
 from pytest_httpx import HTTPXMock
 
 from kfinance.client.models.response_models import PostResponse
-from kfinance.conftest import SPGI_ID_TRIPLE
+from kfinance.conftest import FAKE_COMPANY_1_ID_TRIPLE, FAKE_COMPANY_2_ID_TRIPLE, SPGI_ID_TRIPLE
 from kfinance.domains.companies.company_models import COMPANY_ID_PREFIX
 from kfinance.domains.line_items.response_notes import (
     FISCAL_PERIOD_WARNING,
@@ -97,7 +97,8 @@ class TestStatements:
         """
 
         expected_resp = GetFinancialStatementFromIdentifiersResp(
-            results={"SPGI": StatementsResp.model_validate(self.statement_resp)},
+            identifier_results={"SPGI": StatementsResp.model_validate(self.statement_resp)},
+            identifier_info={"SPGI": SPGI_ID_TRIPLE},
             errors=[
                 "No identification triple found for the provided identifier: NON-EXISTENT of type: ticker"
             ],
@@ -130,7 +131,8 @@ class TestStatements:
         )
 
         expected_resp = GetFinancialStatementFromIdentifiersResp(
-            results={},
+            identifier_results={},
+            identifier_info={"SPGI": SPGI_ID_TRIPLE},
             errors=["SPGI: No results found."],
             notes=[FISCAL_PERIOD_WARNING, FISCAL_YEAR_TERMINOLOGY_WARNING],
         )
@@ -175,10 +177,11 @@ class TestStatements:
         )
 
         expected_response = GetFinancialStatementFromIdentifiersResp(
-            results={
+            identifier_results={
                 "C_1": expected_single_company_response,
                 "C_2": expected_single_company_response,
             },
+            identifier_info={"C_1": FAKE_COMPANY_1_ID_TRIPLE, "C_2": FAKE_COMPANY_2_ID_TRIPLE},
             notes=[FISCAL_PERIOD_WARNING, FISCAL_YEAR_TERMINOLOGY_WARNING],
         )
 
@@ -201,7 +204,8 @@ class TestStatements:
         """
 
         expected_resp = GetFinancialStatementFromIdentifiersResp(
-            results={},
+            identifier_results={},
+            identifier_info={},
             errors=[
                 "No identification triple found for the provided identifier:"
                 " NON-EXISTENT of type: ticker"