+++ kfinance/integrations/tool_calling/tests/test_tool_calling_models.py
@@ -9,7 +9,7 @@ import pytest
 from pytest_httpx import HTTPXMock
 
 from kfinance.client.kfinance import Client
-from kfinance.conftest import SPGI_COMPANY_ID, SPGI_ID_TRIPLE
+from kfinance.conftest import SPGI_COMPANY_ID, SPGI_ID_TRIPLE, SPGI_TICKER
 from kfinance.domains.business_relationships.business_relationship_models import (
     BusinessRelationshipType,
 )
@@ -17,19 +17,26 @@ from kfinance.domains.business_relationships.business_relationship_tools import 
     GetBusinessRelationshipFromIdentifiers,
     GetBusinessRelationshipFromIdentifiersArgs,
 )
-from kfinance.domains.companies.company_models import COMPANY_ID_PREFIX
+from kfinance.domains.companies.company_models import (
+    COMPANY_ID_PREFIX,
+    IdentificationTripleWithCompanyInfo,
+)
 from kfinance.domains.companies.company_tools import (
     GetInfoFromIdentifiers,
     GetInfoFromIdentifiersResp,
 )
-from kfinance.integrations.tool_calling.tool_calling_models import ValidQuarter
+from kfinance.integrations.tool_calling.tool_calling_models import (
+    IdentifierInfoWithResult,
+    ToolRespWithIdInfoAndErrors,
+    ValidQuarter,
+)
 
 
 class TestGetEndpointsFromToolCallsWithGrounding:
     @pytest.mark.asyncio
     async def test_get_info_from_identifier_with_grounding(
         self, mock_client: Client, httpx_mock: HTTPXMock
-    ):
+    ) -> None:
         """
         GIVEN a KfinanceTool tool
         WHEN we run the tool with `run_with_grounding`
@@ -41,16 +48,18 @@ class TestGetEndpointsFromToolCallsWithGrounding:
             "name": "S&P Global Inc.",
             "status": "Operating",
             "company_id": f"{COMPANY_ID_PREFIX}{SPGI_COMPANY_ID}",
+            "ticker": SPGI_TICKER,
         }
         resp_endpoint = [
             "https://kfinance.kensho.com/api/v1/ids",
             "https://kfinance.kensho.com/api/v1/info/21719",
         ]
         expected_resp = {
-            "data": GetInfoFromIdentifiersResp.model_validate({"results": {"SPGI": resp_data}}),
+            "data": GetInfoFromIdentifiersResp(results={"SPGI": resp_data}),
             "endpoint_urls": resp_endpoint,
         }
         del resp_data["company_id"]
+        del resp_data["ticker"]
 
         # Mock the /ids endpoint
         httpx_mock.add_response(
@@ -103,7 +112,7 @@ class TestValidQuarter:
 
 
 class TestRunSyncAndAsync:
-    def test_run_sync_and_async(self, add_spgi_supplier_mock_resp: None, httpx_client: Any):
+    def test_run_sync_and_async(self, add_spgi_supplier_mock_resp: None, httpx_client: Any) -> None:
         """
         GIVEN a sync environment with sync and async clients (via asyncio.run)
         WHEN requests are made with both sync and async clients in a sync
@@ -125,12 +134,12 @@ class TestRunSyncAndAsync:
             datetime(2100, 1, 1).timestamp()
         )
 
-        def run_sync():
+        def run_sync() -> Any:
             tool = GetBusinessRelationshipFromIdentifiers(kfinance_client=sync_client)
             sync_resp = tool.run(args.model_dump(mode="json"))
             return sync_resp
 
-        async def run_async_twice():
+        async def run_async_twice() -> Any:
             tool = GetBusinessRelationshipFromIdentifiers(kfinance_client=async_client)
             async_resp1 = await tool.ainvoke(args.model_dump(mode="json"))
             async_resp2 = await tool.ainvoke(args.model_dump(mode="json"))
@@ -148,3 +157,67 @@ class TestRunSyncAndAsync:
         sync_res4 = run_sync()
 
         assert sync_res1 == sync_res2 == sync_res3 == sync_res4 == async_res1 == async_res2
+
+
+class TestToolResp:
+    @pytest.mark.parametrize(
+        "identifier_results,identifier_info,expected_results",
+        [
+            pytest.param({}, {}, {}, id="no ids"),
+            pytest.param(
+                {},
+                {
+                    "COMP": IdentificationTripleWithCompanyInfo(
+                        company_id=10,
+                        security_id=None,
+                        trading_item_id=None,
+                        company_name="Company",
+                        ticker=None,
+                        country=None,
+                    )
+                },
+                {},
+                id="no id results, but some identifier info. there should still be no output results.",
+            ),
+            pytest.param(
+                {"COMP": 5, "OTHER": 7},
+                {
+                    "COMP": IdentificationTripleWithCompanyInfo(
+                        company_id=10,
+                        security_id=None,
+                        trading_item_id=None,
+                        company_name="Company",
+                        ticker=None,
+                        country=None,
+                    ),
+                    "OTHER": IdentificationTripleWithCompanyInfo(
+                        company_id=20,
+                        security_id=1,
+                        trading_item_id=1,
+                        company_name="Other Company",
+                        ticker="EX:OTH",
+                        country="SWE",
+                    ),
+                },
+                {
+                    "COMP": IdentifierInfoWithResult(
+                        company_name="Company", ticker=None, country=None, data=5
+                    ),
+                    "OTHER": IdentifierInfoWithResult(
+                        company_name="Other Company", ticker="EX:OTH", country="SWE", data=7
+                    ),
+                },
+                id="multiple id results",
+            ),
+        ],
+    )
+    def test_tool_resp_with_id_info(
+        self,
+        identifier_results: dict[str, int],
+        identifier_info: dict[str, IdentificationTripleWithCompanyInfo],
+        expected_results: dict[str, IdentifierInfoWithResult[int]],
+    ) -> None:
+        tool_resp = ToolRespWithIdInfoAndErrors[int](
+            identifier_results=identifier_results, identifier_info=identifier_info
+        )
+        assert expected_results == tool_resp.results