+++ kfinance/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v6.0.0
+- Update tool call results to include company information - in particular, 
+  company name, ticker, and country
+- Breaking changes:
+  - The result type of identifier-based tool calls has been modified.
+
 ## v5.1.4
 - Update tools to gracefully handle more errors